pyproject.toml
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "dusken.settings.dev"
# FIXME: https://github.com/django-extensions/django-extensions/pull/1741
filterwarnings = "ignore::PendingDeprecationWarning:django.apps.registry.*"
addopts = "--disable-socket"

[tool.coverage.run]
omit = [
    "dusken/migrations/*",
]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "D",               # pydocstyle: Allow undocumented code
  "ANN",             # flake8-annotations: Allow untyped code
  "TD", "FIX",       # Allow TODOs and FIXMEs
  "RUF012",          # FIXME: Fix later https://docs.astral.sh/ruff/rules/mutable-class-default/
  "FBT002",          # Allow boolean args
  "G004",            # Allow logging using f-string
  "EM101", "EM102",  # Allow string in exceptions (prioritice code readability vs exception time readability)
  "TRY003",          # Allow long exception messages
  "E501",            # Allow long lines
  "PGH003",          # FIXME: Improve typing
  "COM812", "ISC001" # Avoid conflicting with ruff formatter
]
[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.per-file-ignores]
# Testing code can deviate a bit more
"dusken/tests/*.py" = ["S101", "S106", "PLR2004"]
"**/test_*.py" = ["S101", "S106", "PLR2004"]
"**/tests.py" = ["S101", "S106", "PLR2004"]
"dusken/apps/tekstmelding/models.py" = ["DJ001"]  # This is autogenerated

[tool.mypy]
disable_error_code = "annotation-unchecked"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main", "strawberry.ext.mypy_plugin"]

# Third party code without type stubs
[[tool.mypy.overrides]]
module = [
    "captcha.*",
    "dj_database_url.*",
    "django_filters.*",
    "django_select2.*",
    "fabric.*",
    "ldapdb.*",
    "mptt.*",
    "oauth2_provider.*",
    "passlib.*",
    "phonenumber_field.*",
    "dusken.settings.local",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "dusken.settings.dev"

[tool.poetry]
name = "galtinn"
version = "0.1.0"
description = ""
authors = ["EDB <edb@neuf.no>"]
license = "MIT"

[tool.poetry.dependencies]
argon2-cffi = "21.3.*"
celery = "5.2.*"
dj-database-url = "0.5.*"
Django = "4.2.*"
django-bootstrap5 = "21.3"
django-countries = "7.6.*"
django-extensions = "3.2.*"
django-filter = "22.1"
django-inline-svg = "0.1.*"
django-ldapdb = { git = "https://github.com/nikolaik/django-ldapdb", rev = "nikolaik/django-42"}
django-mptt = "0.13.*"
django-phonenumber-field = "6.4.*"
django-recaptcha = "3.0.*"
django-redis = "5.2.*"
django-select2 = "7.10.*"
django-signal-disabler = "0.1.*"
djangorestframework = "3.15.*"
fabric = "2.7.*"
gunicorn = "20.1.*"
Markdown = "3.4.*"
mysqlclient = "2.2.*"
passlib = "1.7.*"
phonenumbers = "8.12.*"
psycopg2 = "2.9.*"
python = "~3.8"
redis = "4.3.*"
requests = "2.28.*"
sentry-sdk = "1.7.*"
strawberry-graphql = "0.224.*"
strawberry-graphql-django = "0.37.*"
stripe = "10.1.*"
django-oauth-toolkit = "^2.3.0"

[tool.poetry.dev-dependencies]
celery-types = "0.13.*"
codecov = "2.1.*"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.0"}
django-upgrade = "^1.16.0"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.0"}
mypy = "1.2.*"
pytest = "*"
pytest-cov = "^5.0.0"
pytest-django = "*"
pytest-only = "*"
pytest-socket = "*"
requests-mock = "1.9.*"
ruff = "0.4.9"
types-requests = "2.28.*"
types-stripe = "^3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
